// src/app/models/userInv.ts
export class userInv {
    iduser: number;
    firstname: string;
    secondname: string;
    surname: string;
    secondsurname: string;
    email: string;
    pass: string;
    phone: string;
    username: string;
    dateborn: string;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
    idrol: number;
    idsucursal: number;
    active: number;
    nombresucursal: string;
}